% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abrigapp.R
\name{abrigapp}
\alias{abrigapp}
\title{demo app 2}
\usage{
abrigapp(con, genelocs)
}
\arguments{
\item{con}{a DBI connection}

\item{genelocs}{a GRanges instance with gene addresses}
}
\description{
demo app 2
}
\note{
Very specialized, just has a few genes, uses specific
field from genelocs argument.
}
\examples{
if (interactive()){
utils::data("gloc_hg19", package = "GGIpack2")
con = DBI::dbConnect(duckdb::duckdb())
abrigapp(con, gloc_hg19)
}
}
