% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{ABRIGresource}
\alias{ABRIGresource}
\title{define constructor for ABRIGresource}
\usage{
ABRIGresource(con, tissue, space = "hg19", pfiles)
}
\arguments{
\item{con}{is a DBI connection (typically duckdb)}

\item{tissue}{character(1)}

\item{space}{character(1) e.g., "hg19"}

\item{pfiles}{list of absolute paths to the data for each tissue.}
}
\value{
ABRIGresource instance for the identified tissue
}
\description{
define constructor for ABRIGresource
}
\examples{
con = DBI::dbConnect(duckdb::duckdb())
options(useFancyQuotes=FALSE)
ll = ABRIGresource( con, "BAL" , pfiles= ABRIGparquet_paths())
print(ll)
DBI::dbDisconnect(con)
}
