% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{checkData}
\alias{checkData}
\title{Takes in a path to a data table and coheres the data into the proper format}
\usage{
checkData(path)
}
\arguments{
\item{path}{The complete path to the DN8 file.}
}
\value{
a data set that can be used to graph in JBrowseR.
}
\description{
Takes in a path to a data table and coheres the data into the proper format
}
\examples{
path = system.file("extdata/Alveolar_Macrophages_IS.MICA_ILMN_3241692.CAU.meta", package="GGIpack2")
head(checkData(path = path))
}
