% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{dorounds}
\alias{dorounds}
\title{Takes in a metaABRIG data table and rounds the SE MAF BETA FDR and statistic to 3 digits. For P (pvalue) that variable is turned into  the scientific notation then rounded to 3 digits.}
\usage{
dorounds(mydf)
}
\arguments{
\item{mydf}{metaABRIG data table}
}
\value{
a metaABRIG data table with rounded values .
}
\description{
Takes in a metaABRIG data table and rounds the SE MAF BETA FDR and statistic to 3 digits. For P (pvalue) that variable is turned into  the scientific notation then rounded to 3 digits.
}
\examples{
con = DBI::dbConnect(duckdb::duckdb())
nn = make_data_frame_from_tissue_and_gene(con, "BAL", "DSP")
dorounds(mydf= nn )
DBI::dbDisconnect(con)
}
