% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens2sym.R
\name{ens2sym}
\alias{ens2sym}
\title{map naked ensembl gene ids to symbols when available in gmap,
otherwise just pass back the ENSG tag}
\usage{
ens2sym(e, gmap)
}
\arguments{
\item{e}{character() ENSG symbols}

\item{gmap}{a character vector with gene symbols as element values and ENSG tags as element names}
}
\description{
map naked ensembl gene ids to symbols when available in gmap,
otherwise just pass back the ENSG tag
}
\examples{
ensg = setup_ens_gnames()
gsyms = names(ensg)
names(gsyms) = as.character(ensg)
head(ens2sym(head(ensg), gsyms))
}
