% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ensg}
\alias{ensg}
\title{A map from symbols to Ensembl IDs}
\format{
data object
}
\usage{
ensg
}
\description{
A map from symbols to Ensembl IDs
}
\note{
this was made primarily to allow
mapping of GTEx v7 molecular_trait_ids to 
symbols.
}
\keyword{datasets}
