% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useosn.R
\name{ggi_gtex_cache}
\alias{ggi_gtex_cache}
\title{add or retrieve selected parquet file from BiocFileCache}
\usage{
ggi_gtex_cache(
  pfname = "lungpl05.parquet",
  cache = BiocFileCache::BiocFileCache()
)
}
\arguments{
\item{pfname}{character(1) name of a parquet file assumed present in Bioc Open Storage Network}

\item{cache}{BiocFileCache-type cache}
}
\description{
add or retrieve selected parquet file from BiocFileCache
}
\examples{
ggi_gtex_cache()
}
