% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formals.R
\docType{class}
\name{ggiResource-class}
\alias{ggiResource-class}
\title{manage a GGI resource which must have start, end, seqnames, space, score
it also can have a tbl element}
\description{
manage a GGI resource which must have start, end, seqnames, space, score
it also can have a tbl element
}
