% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gloc_hg19}
\alias{gloc_hg19}
\title{A instance of GenomicRanges human genome 19}
\format{
data object
}
\usage{
gloc_hg19
}
\description{
A instance of GenomicRanges human genome 19
}
\keyword{datasets}
