% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtexapp.R
\name{gtexapp}
\alias{gtexapp}
\title{run the app with GTEx lung and whole blood}
\usage{
gtexapp()
}
\description{
run the app with GTEx lung and whole blood
}
\examples{
oldask = options()$ask
if (interactive()) {
  options(ask=FALSE)
  gtexapp()
}
options(ask=oldask)
}
