% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formals.R
\name{GTExresource}
\alias{GTExresource}
\title{constructor for GTEx examples}
\usage{
GTExresource(con, space = "hg19", tisstag, pfile)
}
\arguments{
\item{con}{DBI connection}

\item{space}{character(1) must indicate build}

\item{tisstag}{character(1) string for tissue name}

\item{pfile}{character(1) path to parquet file}
}
\description{
constructor for GTEx examples
}
\note{
'score' is p-value
}
\examples{
fo = ggi_gtex_cache("lungpl05.parquet")
if (nchar(fo)>0) {
  lungpa = fo
  con = DBI::dbConnect(duckdb::duckdb())
  lu = GTExresource(con, tisstag="lung", pfile=lungpa)
  lu
# please do dbDisconnect(con) when done with 'lu'
}
}
