% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{make_data_frame_from_tissue_and_gene}
\alias{make_data_frame_from_tissue_and_gene}
\title{this makes a data.frame}
\usage{
make_data_frame_from_tissue_and_gene(con, tissue, gene)
}
\arguments{
\item{con}{is a DBI connection (typically duckdb)}

\item{tissue}{character(1)}

\item{gene}{character(1)}
}
\value{
a data frame filters by tissue and gene.
}
\description{
this makes a data.frame
}
\examples{
con = DBI::dbConnect(duckdb::duckdb())
nn = make_data_frame_from_tissue_and_gene(con, "BAL", "DSP")
DBI::dbDisconnect(con)
}
