% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{newensg}
\alias{newensg}
\title{map from symbols to v79 Ensembl IDs}
\format{
named character vector
}
\usage{
newensg
}
\description{
map from symbols to v79 Ensembl IDs
}
\note{
previous 'ensg' vector had lots of missing
symbol mappings.  This combines a more recent
ensembl catalog with 'obsolete' mappings that
were present in ensg.
}
\keyword{datasets}
