% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_gnames.R
\name{setup_ens_gnames}
\alias{setup_ens_gnames}
\title{use newensg data element to provide a mapping between symbol and ENSG id}
\usage{
setup_ens_gnames()
}
\description{
use newensg data element to provide a mapping between symbol and ENSG id
}
\examples{
head(setup_ens_gnames())
}
