% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formals.R
\name{show,GTExresource-method}
\alias{show,GTExresource-method}
\title{present information about a ggiResource tailored to GTEx v7 summaries}
\usage{
\S4method{show}{GTExresource}(object)
}
\description{
present information about a ggiResource tailored to GTEx v7 summaries
}
