% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyapp.R
\name{tinyapp2}
\alias{tinyapp2}
\title{demo app 2}
\usage{
tinyapp2(con, genelocs)
}
\arguments{
\item{con}{a DBI connection}

\item{genelocs}{a GRanges instance with gene addresses}
}
\description{
demo app 2
}
\note{
Very specialized, just has a few genes, uses specific
field from genelocs argument.
}
\examples{
if (interactive()){
utils::data("gloc_hg19", package = "GGIpack")
con = DBI::dbConnect(duckdb::duckdb())
tinyapp2(con, gloc_hg19)
}
}
