% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toGRanges.R
\name{toGRanges}
\alias{toGRanges}
\title{convert the tabular information from a ggiResource instance into GRanges}
\usage{
toGRanges(tbl, genome = NA, maxn.msg = 50000)
}
\arguments{
\item{tbl}{resource with seqnames, start, end}

\item{genome}{character(1) should come from resource 'space' element}

\item{maxn.msg}{numeric(1) if number of records for conversion to GRanges exceeds this, a message is given}
}
\description{
convert the tabular information from a ggiResource instance into GRanges
}
