% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topscores_by.R
\name{topscores_by}
\alias{topscores_by}
\title{get scores of selected types from a GGIresource instance}
\usage{
topscores_by(res, scope = "chr", val, n = 100)
}
\arguments{
\item{res}{GGIresource instance}

\item{scope}{character(1) one of c("chr", "gene", "rsid")}

\item{val}{a value in the selected `scope`}

\item{n}{numeric(1) used to filter results through `head()`}
}
\description{
get scores of selected types from a GGIresource instance
}
\examples{
# we rely on duckdb for query resolution
blex = system.file("parquet", "testbl.parquet", package="GGIpack2")
pp = sprintf("read_parquet(\%s)", sQuote(blex, q=FALSE))
con = DBI::dbConnect(duckdb::duckdb())
tb = dplyr::tbl(con, pp)
nres = new("GTExresource", space = "hg19", tbl = tb)
topscores_by(nres, scope="chr", val="1", n=5)
topscores_by(nres, scope="gene", val="FO538757.2", n=5)
topscores_by(nres, scope="rsid", val="rs368811019", n=5)
DBI::dbDisconnect(con)
}
