% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abrigapp2.R
\name{abrigapp2}
\alias{abrigapp2}
\title{run the app with two abrig resources}
\usage{
abrigapp2()
}
\description{
run the app with two abrig resources
}
\examples{
oldask = options()$ask
if (interactive()) {
  options(ask=FALSE)
  abrigapp2()
}
options(ask=oldask)
}
