% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formals.R
\docType{class}
\name{GTExresource-class}
\alias{GTExresource-class}
\title{GTEx resource is tallored to the  wholeblpl05.parquet and lungpl05.parquet  resources.}
\description{
GTEx resource is tallored to the  wholeblpl05.parquet and lungpl05.parquet  resources.

GTEx resource is tallored to the  wholeblpl05.parquet and lungpl05.parquet  resources.
}
